Refactor users table action buttons
